Locations (#35)
* added delete command for locations

* location changes

* Added messages if the command is done wrong and fixed some minor issues

---------

+++ src/main/java/com/rosekingdom/rosekingdom/Locations/subCommands/delete.java
@@ -3,6 +3,8 @@ package com.rosekingdom.rosekingdom.Locations.subCommands;
 import com.rosekingdom.rosekingdom.Core.CommandManager.subCommandRK;
 import com.rosekingdom.rosekingdom.Core.Database.Main_Statements.UserStatement;
 import com.rosekingdom.rosekingdom.Locations.Statements.LocationStatement;
+import net.kyori.adventure.text.Component;
+import net.kyori.adventure.text.format.TextColor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
@@ -10,6 +12,7 @@ public class delete extends subCommandRK {
         public delete(int arg){
             super(arg);
             setName("delete");
+            addAlias("remove");
         }
 
         @Override
@@ -18,10 +21,11 @@ public class delete extends subCommandRK {
                 return;
             }
             int id = UserStatement.getId(player.getUniqueId());
-            if(args.length==2){
+            if(args.length==2 && LocationStatement.exists(id, args[1])){
                 LocationStatement.deleteLocation(id, args[1]);
+                player.sendMessage(Component.text("Location " + args[1] + " was deleted!", TextColor.fromHexString("#6be649")));
+            }else{
+                player.sendMessage(Component.text("Missing arguments!", TextColor.fromHexString("#e30000")));
             }
-
         }
-
     }
